refactor: move menu domain to store package
