+++ src/main/kotlin/com/carava/carwash/store/dto/CreateMenuRequestDto.kt
@@ -1,9 +1,8 @@
-package com.carava.carwash.menu.dto
+package com.carava.carwash.store.dto
 
 import java.math.BigDecimal
 
 data class CreateMenuRequestDto(
-    val storeId: Long,
     val name: String,
     val price: BigDecimal,
 )
